+++ common/src/main/java/de/maxhenkel/voicechat/voice/server/GroupManager.java
@@ -14,7 +14,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
 
 public class GroupManager {
 
@@ -85,14 +84,15 @@ public class GroupManager {
 
         PlayerStateManager manager = getStates();
         manager.setGroup(player.server, player, null);
+        NetManager.sendToClient(player, new JoinedGroupPacket(null, false));
 
         cleanEmptyGroups();
     }
 
     public void cleanEmptyGroups() {
         PlayerStateManager manager = getStates();
-        List<UUID> usedGroups = manager.getStates().stream().filter(PlayerState::hasGroup).map(state -> state.getGroup().getId()).distinct().collect(Collectors.toList());
-        List<UUID> groupsToRemove = groups.keySet().stream().filter(uuid -> !usedGroups.contains(uuid)).collect(Collectors.toList());
+        List<UUID> usedGroups = manager.getStates().stream().filter(PlayerState::hasGroup).map(state -> state.getGroup().getId()).distinct().toList();
+        List<UUID> groupsToRemove = groups.keySet().stream().filter(uuid -> !usedGroups.contains(uuid)).toList();
         for (UUID uuid : groupsToRemove) {
             groups.remove(uuid);
         }