+++ fabric/changelog.md
@@ -1,1 +1,1 @@
-- Fixed icon sizes
+- Fixed leaving group not working